requirements.txt
pandas
requests
colorlog
python-dotenv
openpyxl